backend/agents/Refactoring/SK/requirements.txt
semantic-kernel
python-a2a
fastapi
uvicorn
a2a-sdk
python-dotenv
